requirements.txt

fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
motor==3.3.2
qdrant-client==1.6.9
sentence-transformers==2.2.2
google-generativeai==0.3.2
Pillow==10.1.0
aiofiles
httpx
torch==2.1.0
torchvision==0.16.0
clip-by-openai==1.0